Refactored product functions
